# specs/001-finance-hub/contracts/api.yaml
openapi: 3.1.0
info:
  title: Finance Hub API
  version: 1.0.0
  description: Self-hosted financial management API
  contact:
    name: Finance Hub
  license:
    name: MIT
servers:
  - url: /api/v1
    description: API v1
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UUID:
      type: string
      format: uuid
...
    Workspace:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
        type:
          type: string
          enum: [personal, family, flatshare, company]
        currency:
          type: string
        role:
          type: string
          enum: [owner, admin, accountant, member, readonly]
        memberCount:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    Transaction:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        accountId:
          $ref: '#/components/schemas/UUID'
        type:
          type: string
          enum: [expense, income, transfer]
        amount:
          type: number
        currency:
          type: string
        date:
          type: string
          format: date
        description:
          type: string
        notes:
          type: string
          nullable: true
        category:
          $ref: '#/components/schemas/CategoryRef'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/TagRef'
        status:
          type: string
          enum: [uncleared, cleared, reconciled]
        documents:
          type: array
          items:
            $ref: '#/components/schemas/DocumentRef'
        createdAt:
          type: string
          format: date-time
...
    CategoryRef:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
        color:
          type: string
        icon:
          type: string
    TagRef:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
        color:
          type: string
    DocumentRef:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        filename:
          type: string
        mimeType:
          type: string

    # Category
...
    ImportPreview:
      type: object
      properties:
        totalRows:
          type: integer
        validRows:
          type: integer
        duplicates:
          type: integer
        errors:
          type: array
          items:
            type: object
            properties:
              row:
                type: integer
              message:
                type: string
        preview:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
          maxItems: 10
    ImportRequest:
      type: object
      required: [accountId, mapping]
      properties:
        accountId:
          $ref: '#/components/schemas/UUID'
        mapping:
          type: object
          properties:
            date:
              type: string
            amount:
              type: string
            description:
              type: string
            category:
              type: string
        skipDuplicates:
          type: boolean
          default: true
        applyRules:
          type: boolean
          default: true
    ImportResult:
      type: object
      properties:
        jobId:
          $ref: '#/components/schemas/UUID'
        status:
          type: string
          enum: [pending, processing, completed, failed]
        imported:
          type: integer
        skipped:
          type: integer
        errors:
          type: integer

    # Settlement (Flatshare)
...
paths:
...
  /workspaces/{workspaceId}:
    parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'

    get:
      tags: [Workspaces]
      summary: Get workspace details
      responses:
        '200':
          description: Workspace details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
...
  /workspaces/{workspaceId}/import/preview:
    parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'

    post:
      tags: [Import]
      summary: Preview CSV import
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                accountId:
                  $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Import preview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportPreview'
  /workspaces/{workspaceId}/import/execute:
    parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'

    post:
      tags: [Import]
      summary: Execute CSV import
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportRequest'
      responses:
        '202':
          description: Import job started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportResult'

  # Reports
...